Struct wait_queue::WaitQueue

pub struct WaitQueue(_);

A queue in which multiple Tasks can wait for other Tasks to notify them.

This can be shared across multiple Tasks by wrapping it in an Arc.

Implementations

impl WaitQueue

pub fn new() -> WaitQueue

Create a new empty WaitQueue.

pub fn with_capacity(initial_capacity: usize) -> WaitQueue

Create a new empty WaitQueue.

pub fn wait(&self) -> Result<(), WaitError>

Puts the current Task to sleep where it blocks on this WaitQueue until it is notified by another Task.

If the Task wakes up spuriously (it is still on the waitqueue), it will be automatically put back to sleep until it is properly woken up. Therefore, there is no need for the caller to check for spurious wakeups.

This function blocks until the Task is woken up through the notify mechanism.

pub fn wait_until<R>(
&self,
condition: &dyn Fn() -> Option<R>
) -> Result<R, WaitError>

Similar to wait, but this function blocks until the given condition closure returns Some(value), and then returns that value inside Ok().

The condition will be executed atomically with respect to the wait queue, which avoids the problem of a waiting task missing a “notify” from another task due to interleaving of instructions that may occur if the condition is checked when the wait queue lock is not held.

pub fn wait_until_mut<R>(
&self,
condition: &mut dyn FnMut() -> Option<R>
) -> Result<R, WaitError>

Similar to wait_until, but this function accepts a condition closure that can mutate its environment (a FnMut).

pub fn notify_one(&self) -> bool

Wake up one random Task that is waiting on this queue.

Return

returns true if a Task was successfully woken up,

returns false if there were no Tasks waiting.

pub fn notify_specific(&self, task_to_wakeup: &TaskRef) -> bool

Wake up a specific Task that is waiting on this queue.

Return

returns true if the given Task was waiting and was woken up,

returns false if there was no such Task waiting.

pub fn notify_all(&self)

Wake up all Tasks that are waiting on this queue.